refactor existing classes to prepare for DML integration testing cases
