+++ src/integrationTest/java/com/mongodb/hibernate/query/AbstractQueryIntegrationTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.mongodb.hibernate.query.select;
+package com.mongodb.hibernate.query;
 
 import static com.mongodb.hibernate.MongoTestAssertions.assertIterableEq;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -36,17 +36,17 @@ import org.junit.jupiter.api.extension.ExtendWith;
 
 @SessionFactory(exportSchema = false)
 @ExtendWith(MongoExtension.class)
-abstract class AbstractSelectionQueryIntegrationTests implements SessionFactoryScopeAware, ServiceRegistryScopeAware {
+public abstract class AbstractQueryIntegrationTests implements SessionFactoryScopeAware, ServiceRegistryScopeAware {
 
     private SessionFactoryScope sessionFactoryScope;
 
     private TestCommandListener testCommandListener;
 
-    SessionFactoryScope getSessionFactoryScope() {
+    protected SessionFactoryScope getSessionFactoryScope() {
         return sessionFactoryScope;
     }
 
-    TestCommandListener getTestCommandListener() {
+    protected TestCommandListener getTestCommandListener() {
         return testCommandListener;
     }
 
@@ -60,7 +60,7 @@ abstract class AbstractSelectionQueryIntegrationTests implements SessionFactoryS
         this.testCommandListener = serviceRegistryScope.getRegistry().requireService(TestCommandListener.class);
     }
 
-    <T> void assertSelectionQuery(
+    protected <T> void assertSelectionQuery(
             String hql,
             Class<T> resultType,
             Consumer<SelectionQuery<T>> queryPostProcessor,
@@ -74,11 +74,12 @@ abstract class AbstractSelectionQueryIntegrationTests implements SessionFactoryS
                 resultList -> assertIterableEq(expectedResultList, resultList));
     }
 
-    <T> void assertSelectionQuery(String hql, Class<T> resultType, String expectedMql, List<T> expectedResultList) {
+    protected <T> void assertSelectionQuery(
+            String hql, Class<T> resultType, String expectedMql, List<T> expectedResultList) {
         assertSelectionQuery(hql, resultType, null, expectedMql, expectedResultList);
     }
 
-    <T> void assertSelectionQuery(
+    protected <T> void assertSelectionQuery(
             String hql,
             Class<T> resultType,
             Consumer<SelectionQuery<T>> queryPostProcessor,
@@ -97,12 +98,12 @@ abstract class AbstractSelectionQueryIntegrationTests implements SessionFactoryS
         });
     }
 
-    <T> void assertSelectionQuery(
+    protected <T> void assertSelectionQuery(
             String hql, Class<T> resultType, String expectedMql, Consumer<List<T>> resultListVerifier) {
         assertSelectionQuery(hql, resultType, null, expectedMql, resultListVerifier);
     }
 
-    <T> void assertSelectQueryFailure(
+    protected <T> void assertSelectQueryFailure(
             String hql,
             Class<T> resultType,
             Consumer<SelectionQuery<T>> queryPostProcessor,
@@ -120,7 +121,7 @@ abstract class AbstractSelectionQueryIntegrationTests implements SessionFactoryS
                 .hasMessage(expectedExceptionMessage, expectedExceptionMessageParameters));
     }
 
-    <T> void assertSelectQueryFailure(
+    protected <T> void assertSelectQueryFailure(
             String hql,
             Class<T> resultType,
             Class<? extends Exception> expectedExceptionType,
@@ -135,7 +136,7 @@ abstract class AbstractSelectionQueryIntegrationTests implements SessionFactoryS
                 expectedExceptionMessageParameters);
     }
 
-    void assertActualCommand(BsonDocument expectedCommand) {
+    protected void assertActualCommand(BsonDocument expectedCommand) {
         var capturedCommands = testCommandListener.getStartedCommands();
 
         assertThat(capturedCommands)
