+++ src/integrationTest/java/com/mongodb/hibernate/query/mutate/DeletionIntegrationTests.java
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2025-present MongoDB, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.mongodb.hibernate.query.mutate;
+
+import com.mongodb.hibernate.query.AbstractQueryIntegrationTests;
+import com.mongodb.hibernate.query.Book;
+import org.hibernate.testing.orm.junit.DomainModel;
+
+@DomainModel(annotatedClasses = Book.class)
+class DeletionIntegrationTests extends AbstractQueryIntegrationTests {}
